+++ src/main/java/edu/infsci2560/MvcConfig.java
@@ -3,6 +3,8 @@ package edu.infsci2560;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @Configuration
 public class MvcConfig extends WebMvcConfigurerAdapter {
@@ -10,11 +12,17 @@ public class MvcConfig extends WebMvcConfigurerAdapter {
     @Override
     public void addViewControllers(ViewControllerRegistry registry) {
         registry.addViewController("/home").setViewName("home");
+        registry.addViewController("/about").setViewName("about");
         registry.addViewController("/").setViewName("home");
         registry.addViewController("/infoArm").setViewName("infoArm");
         registry.addViewController("/login").setViewName("login");
         registry.addViewController("/files").setViewName("profiles");
         registry.addViewController("/certifications").setViewName("certificationsUpdate");
     }
+    @Bean
+        public BCryptPasswordEncoder passwordEncoder() {
+            BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+            return bCryptPasswordEncoder;
+        }
 
 }