+++ src/main/java/edu/infsci2560/WebSecurityConfig.java
@@ -3,31 +3,105 @@ package edu.infsci2560;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
+import javax.sql.DataSource;
 
 @Configuration
 @EnableWebSecurity
 public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+    @Autowired
+	private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+    @Autowired
+    private DataSource dataSource;
+
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http
             .authorizeRequests()
-                .antMatchers("/", "/index", "/about.html").permitAll()
+                .antMatchers("/", "/index", "/about", "/home", "/**",  "/login", "/registration").permitAll() // "/**"
+                .antMatchers("/admin/**").hasAuthority("ADMIN")
                 .anyRequest().authenticated()
                 .and()
+
             .formLogin()
                 .loginPage("/login")
+                .failureUrl("/login?error=true")
+                .defaultSuccessUrl("/")
+                .usernameParameter("username")
+                .passwordParameter("password")
+                // .permitAll()
+                .and()
+
+            .logout()
+                .logoutRequestMatcher(new AntPathRequestMatcher("/logout"))
+                .logoutSuccessUrl("/login")
                 .permitAll()
                 .and()
-            .logout()
-                .permitAll();
+
+            .exceptionHandling()
+                .accessDeniedPage("/access-denied");
     }
+    @Override
+	public void configure(WebSecurity web) throws Exception {
+        
+	    web
+	       .ignoring()
+	       .antMatchers("/resources/**", "/static/**", "/css/**", "/js/**");
+	}
 
     @Autowired
     public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
+        
+        auth
+			.jdbcAuthentication()
+				.usersByUsernameQuery("select email, password, active from users where email=?")
+				.authoritiesByUsernameQuery("select u.email, ur.roles from users u inner join user_roles ur on(u.id=ur.user_id) where u.email=?")
+				.dataSource(dataSource)
+				.passwordEncoder(bCryptPasswordEncoder);
         auth.inMemoryAuthentication().withUser("user").password("password").roles("USER");
-        auth.inMemoryAuthentication().withUser("admin").password("pass").roles("USER");
+        // auth.inMemoryAuthentication().withUser("admin").password("pass").roles("ADMIN");
     }
 }
+
+
+/////////////////////////////////////old configuration
+// package edu.infsci2560;
+
+// import org.springframework.beans.factory.annotation.Autowired;
+// import org.springframework.context.annotation.Configuration;
+// import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+// import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+// import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+// import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+
+// @Configuration
+// @EnableWebSecurity
+// public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+//     @Override
+//     protected void configure(HttpSecurity http) throws Exception {
+//         http
+//             .authorizeRequests()
+//                 .antMatchers("/", "/index", "/about", "/home", "/**").permitAll() // "/**"
+//                 .anyRequest().authenticated()
+//                 .and()
+//             .formLogin()
+//                 .loginPage("/login")
+//                 .permitAll()
+//                 .and()
+//             .logout()
+//                 .permitAll();
+                
+//     }
+
+//     @Autowired
+//     public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
+//         auth.inMemoryAuthentication().withUser("user").password("password").roles("USER");
+//         auth.inMemoryAuthentication().withUser("admin").password("pass").roles("ADMIN");
+//     }
+// }