+++ src/main/java/edu/infsci2560/controllers/CourseController.java
@@ -19,6 +19,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.bind.annotation.PathVariable;
+
 
 /**
  *
@@ -38,5 +40,30 @@ public class CourseController {
     public ModelAndView create(@ModelAttribute @Valid Course course, BindingResult result) {
         repository.save(course);
         return new ModelAndView("courses", "courses", repository.findAll());
+    }
+
+    
+    @RequestMapping(value = "courses/delete/{id}", method = RequestMethod.DELETE, consumes="application/x-www-form-urlencoded", produces = "application/json")
+    public ModelAndView delete(@PathVariable("id") Long id) {
+        // repository.delete(repository.findOne(id));
+        // @ModelAttribute;
+        repository.deleteById(id);
+        // return new ModelAndView("infoArms", "infoArms", repository.findAll());
+        return new ModelAndView("redirect:/courses");
+    }
+
+    @RequestMapping(value = "coursesUpdate/update/{id}", method = RequestMethod.GET)
+    public ModelAndView getRecord(@PathVariable("id") Long id) {
+  
+        Course course = repository.findById(id);
+        ModelAndView modelAndView = new ModelAndView("coursesUpdate");
+        modelAndView.addObject("course", course);
+        return modelAndView;
+    }
+
+    @RequestMapping(value = "coursesUpdate/update/{id}" , method =  RequestMethod.PUT)
+    public ModelAndView create(@PathVariable("id") Long id, @ModelAttribute @Valid Course course, BindingResult result) {
+        repository.save(course);
+        return new ModelAndView("redirect:/courses"); // inforArms template name
     }
 }