+++ src/main/java/edu/infsci2560/models/InfoArm.java
@@ -23,37 +23,43 @@ public class InfoArm {
 
     private static final long serialVersionUID = 1L;
 
-    public enum TicketType {
-        Network_Security,
-        Software_Security,
-        Hardware_Security,
-        Information_Policy_Security
-    }
+    // public enum TicketType {
+    //     Network_Security,
+    //     Software_Security,
+    //     Hardware_Security,
+    //     Information_Policy_Security
+    // }
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     protected Long id;
     protected String ticketTitle;
-    protected TicketType ticketType;
+    // protected TicketType ticketType;
+    protected String ticketType;
     @OneToOne(targetEntity=Orgnization.class, mappedBy="infoArm", cascade=CascadeType.ALL)
     protected Orgnization orgnization;  // orgnization details
+    protected String responseTicket;
 
     public InfoArm() {
         this.id = Long.MAX_VALUE;
         this.ticketTitle = null;
-        this.ticketType = TicketType.Network_Security;
+        // this.ticketType = TicketType.Network_Security;
+        this.ticketType = "Undefined";
+        this.responseTicket = "Not yet";
     }
 
-    public InfoArm(String ticketTitle, TicketType ticketType) {
+    // public InfoArm(String ticketTitle, TicketType ticketType) {
+    public InfoArm(String ticketTitle, String ticketType, String responseTicket) {
         this.ticketTitle = ticketTitle;
         this.ticketType = ticketType;
+        this.responseTicket = responseTicket;
     }
 
     @Override
     public String toString() {
         String var = String.format(
-            "InfoArm[id=%d, ticketTitle='%s',ticketType='%s', orgnization='%s']%n", this.id, this.ticketTitle,
-             this.ticketType, this.orgnization.toString());
+            "InfoArm[id=%d, ticketTitle='%s',ticketType='%s', responseTicket='%s', orgnization='%s']%n", this.id, this.ticketTitle,
+             this.ticketType, this.responseTicket, this.orgnization.toString());
         /*return "[ id=" + this.id + ", ticketTitle=" + this.ticketTitle + ", ticketType=" + this.ticketType + " ]";*/
         return var;
     }
@@ -85,14 +91,18 @@ public class InfoArm {
     /**
      * @return the TicketType
      */
-    public TicketType getTicketType() {
+     
+    // public TicketType getTicketType() {
+    public String getTicketType() {
         return ticketType;
     }
 
     /**
      * @param ticketType the ticketType to set
      */
-    public void setTicketType(TicketType ticketType) {
+
+    // public void setTicketType(TicketType ticketType) {
+    public void setTicketType(String ticketType) {
         this.ticketType = ticketType;
     }
 
@@ -118,4 +128,17 @@ public class InfoArm {
         this.id = id;
     }
 
+    /**
+     * @return the responseTicket
+     */
+    public String getResponseTicket() {
+        return responseTicket;
+    }
+
+    /**
+     * @param responseTicket the responseTicket to set
+     */
+    public void setResponseTicket(String responseTicket) {
+        this.responseTicket = responseTicket;
+    }
 }