+++ src/main/java/edu/infsci2560/models/Orgnization.java
@@ -21,7 +21,7 @@ import org.apache.commons.lang3.builder.HashCodeBuilder;
 public class Orgnization {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     protected Long id;
     protected String orgnizationType;
     protected String numberOfEmpolyee;