+++ src/main/java/edu/infsci2560/services/InfoArmService.java
@@ -6,7 +6,7 @@
 package edu.infsci2560.services;
 
 import edu.infsci2560.models.InfoArm;
-import edu.infsci2560.models.InfoArm.TicketType;
+// import edu.infsci2560.models.InfoArm.TicketType;
 import edu.infsci2560.repositories.InfoArmRepository;
 import java.util.ArrayList;
 import java.util.HashMap;