+++ src/main/java/edu/infsci2560/services/UsersService.java
@@ -1,0 +1,53 @@
+package edu.infsci2560.services;
+
+import edu.infsci2560.models.User;
+import edu.infsci2560.repositories.UserRepository;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.servlet.ModelAndView;
+
+@RestController
+@RequestMapping("/public/api/users")
+public class UsersService {
+
+    @Autowired
+    private UserRepository repository;
+
+    @RequestMapping(value = "/profile", method = RequestMethod.GET)
+    public ModelAndView profile() {
+
+        User user = repository.findById(1L);
+
+        return new ModelAndView("ajax/profile", "user", user);
+    }
+
+    @RequestMapping(method = RequestMethod.GET, produces = "application/json")
+    public ResponseEntity<Iterable<User>> list() {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.findAll(), headers, HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/{id}", method = RequestMethod.GET, produces = "application/json")
+    public ResponseEntity<User> list(@PathVariable("id") Long id) {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.findOne(id), headers, HttpStatus.OK);
+    }
+
+    @RequestMapping(method = RequestMethod.POST, consumes="application/json", produces = "application/json")
+    public ResponseEntity<User> create(@RequestBody User user) {
+        HttpHeaders headers = new HttpHeaders();
+        return new ResponseEntity<>(repository.save(user), headers, HttpStatus.OK);
+    }
+}